requirements.txt
numpy~=1.22.3
configspace~=0.5.0
naslib~=0.1.0
dehb~=0.0.2
fvcore~=0.1.5.post20210630
torch~=1.9.0
torchvision~=0.10.0
scikit-learn~=1.0.2
scipy~=1.8.1
termcolor~=1.1.0
pillow~=9.0.1
scikit-image~=0.19.2
matplotlib~=3.5.1
grakel~=0.1.8
networkx~=2.7.1
six~=1.16.0
pybnn~=0.0.5
emcee~=3.1.0
lightgbm~=3.2.1
ngboost~=0.3.11
xgboost~=1.4.2
tensorwatch~=0.9.1
pyyaml~=5.4.1
setuptools~=62.6.0
deepcave~=1.0
smac~=1.3.4
distributed~=2022.6.1
loguru~=0.6.0
pytest~=6.2.4
requests~=2.28.0
pandas~=1.3.5
jsonlines~=3.0.0
dash~=2.0.0
plotly~=5.9.0
rq~=1.10.1
pyrfr~=0.8.3
tqdm~=4.64.0
redis~=4.3.4